agentic-day2/requirements.txt
langchain
langchain-openai
langchain-core
python-dotenv
openai
langgraph
